fix(ci): resolve GitHub Actions failures
Fix all CI job failures identified in recent workflow runs:

1. Remove unnecessary cast in graph_coloring algorithm
   - Remove redundant `as i32` cast on saturation[i] which is already i32
   - Keep necessary cast on degree[i] from u32 to i32
   - Fixes clippy::unnecessary_cast warning

2. Improve LocalStack health check reliability
   - Use curl -sf to check HTTP status instead of grepping response
   - Increase timeout from 60s to 120s for better reliability
   - Add progress logging to track health check status

3. Fix documentation build collision
   - Exclude uni-python package from workspace documentation
   - Resolves library name conflict between Python bindings and Rust crate

These fixes address test job, cloud integration tests, documentation build,
and Python bindings build failures.

+++ crates/uni-algo/src/algo/algorithms/graph_coloring.rs
@@ -76,7 +76,7 @@ impl Algorithm for GraphColoring {
 
             for i in 0..n {
                 if colors[i].is_none() {
-                    let sat = saturation[i] as i32;
+                    let sat = saturation[i];
 
                     let deg = degree[i] as i32;
